@@ -6,11 +6,11 @@
 
     <groupId>org.appxi</groupId>
     <artifactId>timeago</artifactId>
-    <version>1.0.0</version>
+    <version>1.1.0</version>
     <description>Pure Java version of marlonlom/timeago</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>14</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
